Restructure the project and its files and folders
