+++ Source/SACircularCountdown/SACircularCountdown.h
@@ -1,0 +1,19 @@
+//
+//  SACircularCountdown.h
+//  SACircularCountdown
+//
+//  Created by Stefan Arambasich on 2/26/18.
+//  Copyright © 2018 Stefan Arambasich. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for SACircularCountdown.
+FOUNDATION_EXPORT double SACircularCountdownVersionNumber;
+
+//! Project version string for SACircularCountdown.
+FOUNDATION_EXPORT const unsigned char SACircularCountdownVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SACircularCountdown/PublicHeader.h>
+
+
